deployed model
Deployed the classiffication model to azure machine learning workspace. Removed all python files for classification, we only use an API call to classify
