--- backend/car_detection/custom_layers/__init__.py
@@ -1,1 +1,0 @@
-# Python Package
