--- backend/car_detection/utils.py
@@ -1,19 +1,0 @@
-import cv2 as cv
-
-from resnet_152 import resnet152_model
-
-
-def load_model():
-    model_weights_path = 'car_detection/models/model.96-0.89.hdf5'
-    img_width, img_height = 224, 224
-    num_channels = 3
-    num_classes = 196
-    model = resnet152_model(img_height, img_width, num_channels, num_classes)
-    model.load_weights(model_weights_path, by_name=True)
-    return model
-
-
-def draw_str(dst, target, s):
-    x, y = target
-    cv.putText(dst, s, (x + 1, y + 1), cv.FONT_HERSHEY_PLAIN, 1.0, (0, 0, 0), thickness=2, lineType=cv.LINE_AA)
-    cv.putText(dst, s, (x, y), cv.FONT_HERSHEY_PLAIN, 1.0, (255, 255, 255), lineType=cv.LINE_AA)
