+++ backend/package.json
@@ -7,6 +7,7 @@
     "express": "^4.16.4",
     "fs-extra": "^8.1.0",
     "multer": "^1.4.1",
+    "numjs": "^0.16.0",
     "path": "^0.12.7",
     "request": "^2.88.0"
   }
